chore: decoupled peak&pop, tabbed and safari at NewsTableViewController
